Check in initial revision of compiler
This is mostly functional, but due to a design flaw does not handle
identifier resolution correctly in nested macros in all cases.

Internal errors are caught and reported.

The compiler will be refactored to preserve the identifier semantics,
and simply use a layer of indirection for resolution, rather than
attempted in-place modification.